* Removido acento do namespace NFe.Classes.Informacoes.Detalhe.Produto Especifico
