+++ NFe.Classes/Informacoes/Detalhe/Produto Especifico/CIDE.cs
@@ -30,64 +30,39 @@
 /* http://www.zeusautomacao.com.br/                                             */
 /* Rua Comendador Francisco josé da Cunha, 111 - Itabaiana - SE - 49500-000     */
 /********************************************************************************/
-namespace NFe.Classes.Informacoes.Detalhe.Produto_Específico
+namespace NFe.Classes.Informacoes.Detalhe.Produto_Especifico
 {
-    public class comb : ProdutoEspecifico
+    public class CIDE
     {
-        private decimal? _pMixGn;
-        private decimal? _qTemp;
+        private decimal _qBcProd;
+        private decimal _vAliqProd;
+        private decimal _vCide;
 
         /// <summary>
-        ///     LA02 - Código de produto da ANP
+        ///     LA08 - BC da CIDE
         /// </summary>
-        public string cProdANP { get; set; }
-
-        /// <summary>
-        ///     LA03 - Percentual de Gás Natural para o produto GLP (cProdANP=210203001)
-        /// </summary>
-        public decimal? pMixGN
+        public decimal qBCProd
         {
-            get { return _pMixGn.Arredondar(4); }
-            set { _pMixGn = value.Arredondar(4); }
+            get { return _qBcProd; }
+            set { _qBcProd = value.Arredondar(4); }
         }
 
         /// <summary>
-        ///     LA04 - Código de autorização / registro do CODIF
+        ///     LA09 - Valor da alíquota da CIDE
         /// </summary>
-        public string CODIF { get; set; }
-
-        /// <summary>
-        ///     LA05 - Quantidade de combustível faturada à temperatura ambiente
-        /// </summary>
-        public decimal? qTemp
+        public decimal vAliqProd
         {
-            get { return _qTemp.Arredondar(4); }
-            set { _qTemp = value.Arredondar(4); }
+            get { return _vAliqProd; }
+            set { _vAliqProd = value.Arredondar(4); }
         }
 
         /// <summary>
-        ///     LA06 - Sigla da UF de consumo
+        ///     LA10 - Valor da CIDE
         /// </summary>
-        public string UFCons { get; set; }
-
-        /// <summary>
-        ///     LA07 - Informações da CIDE
-        /// </summary>
-        public CIDE CIDE { get; set; }
-
-        /// <summary>
-        /// LA11 - Informações do grupo de “encerrante”
-        /// </summary>
-        public encerrante encerrante { get; set; }
-
-        public bool ShouldSerializepMixGN()
+        public decimal vCIDE
         {
-            return pMixGN.HasValue;
-        }
-
-        public bool ShouldSerializeqTemp()
-        {
-            return qTemp.HasValue;
+            get { return _vCide; }
+            set { _vCide = value.Arredondar(2); }
         }
     }
 }
