+++ NFe.Classes/Informacoes/Detalhe/Produto Especifico/comb.cs
@@ -30,44 +30,64 @@
 /* http://www.zeusautomacao.com.br/                                             */
 /* Rua Comendador Francisco josé da Cunha, 111 - Itabaiana - SE - 49500-000     */
 /********************************************************************************/
-namespace NFe.Classes.Informacoes.Detalhe.Produto_Específico
+namespace NFe.Classes.Informacoes.Detalhe.Produto_Especifico
 {
-    public class med : ProdutoEspecifico
+    public class comb : ProdutoEspecifico
     {
-        private decimal _qLote;
-        private decimal _vPmc;
+        private decimal? _pMixGn;
+        private decimal? _qTemp;
 
         /// <summary>
-        ///     K02 - Número do Lote de medicamentos ou de matérias-primas farmacêuticas
+        ///     LA02 - Código de produto da ANP
         /// </summary>
-        public string nLote { get; set; }
+        public string cProdANP { get; set; }
 
         /// <summary>
-        ///     K03 - Quantidade de produto no Lote de medicamentos ou de matérias-primas farmacêuticas
+        ///     LA03 - Percentual de Gás Natural para o produto GLP (cProdANP=210203001)
         /// </summary>
-        public decimal qLote
+        public decimal? pMixGN
         {
-            get { return _qLote; }
-            set { _qLote = value.Arredondar(3); }
+            get { return _pMixGn.Arredondar(4); }
+            set { _pMixGn = value.Arredondar(4); }
         }
 
         /// <summary>
-        ///     K04 - Data de fabricação. Formato: “AAAA-MM-DD”
+        ///     LA04 - Código de autorização / registro do CODIF
         /// </summary>
-        public string dFab { get; set; }
+        public string CODIF { get; set; }
 
         /// <summary>
-        ///     K05 - Data de validade. Formato: “AAAA-MM-DD”
+        ///     LA05 - Quantidade de combustível faturada à temperatura ambiente
         /// </summary>
-        public string dVal { get; set; }
-
-        /// <summary>
-        ///     K06 - Preço máximo consumidor
-        /// </summary>
-        public decimal vPMC
+        public decimal? qTemp
         {
-            get { return _vPmc; }
-            set { _vPmc = value.Arredondar(2); }
+            get { return _qTemp.Arredondar(4); }
+            set { _qTemp = value.Arredondar(4); }
+        }
+
+        /// <summary>
+        ///     LA06 - Sigla da UF de consumo
+        /// </summary>
+        public string UFCons { get; set; }
+
+        /// <summary>
+        ///     LA07 - Informações da CIDE
+        /// </summary>
+        public CIDE CIDE { get; set; }
+
+        /// <summary>
+        /// LA11 - Informações do grupo de “encerrante”
+        /// </summary>
+        public encerrante encerrante { get; set; }
+
+        public bool ShouldSerializepMixGN()
+        {
+            return pMixGN.HasValue;
+        }
+
+        public bool ShouldSerializeqTemp()
+        {
+            return qTemp.HasValue;
         }
     }
 }
