+++ NFe.Classes/Informacoes/Detalhe/Produto Especifico/med.cs
@@ -30,48 +30,44 @@
 /* http://www.zeusautomacao.com.br/                                             */
 /* Rua Comendador Francisco josé da Cunha, 111 - Itabaiana - SE - 49500-000     */
 /********************************************************************************/
-namespace NFe.Classes.Informacoes.Detalhe.Produto_Específico
+namespace NFe.Classes.Informacoes.Detalhe.Produto_Especifico
 {
-    public class encerrante
+    public class med : ProdutoEspecifico
     {
-        private decimal _vEncIni;
-        private decimal _vEncFin;
+        private decimal _qLote;
+        private decimal _vPmc;
 
         /// <summary>
-        /// LA12 - Número de identificação do bico utilizado no abastecimento
+        ///     K02 - Número do Lote de medicamentos ou de matérias-primas farmacêuticas
         /// </summary>
-        public int nBico { get; set; }
+        public string nLote { get; set; }
 
         /// <summary>
-        /// LA13 - Número de identificação da bomba ao qual o bico está interligado
+        ///     K03 - Quantidade de produto no Lote de medicamentos ou de matérias-primas farmacêuticas
         /// </summary>
-        public int? nBomba { get; set; }
-        public bool ShouldSerializenBomba()
+        public decimal qLote
         {
-            return nBomba.HasValue;
+            get { return _qLote; }
+            set { _qLote = value.Arredondar(3); }
         }
 
         /// <summary>
-        /// LA14 - Número de identificação do tanque ao qual o bico está interligado
+        ///     K04 - Data de fabricação. Formato: “AAAA-MM-DD”
         /// </summary>
-        public int nTanque { get; set; }
+        public string dFab { get; set; }
 
         /// <summary>
-        /// LA15 - Valor do Encerrante no início do abastecimento
+        ///     K05 - Data de validade. Formato: “AAAA-MM-DD”
         /// </summary>
-        public decimal vEncIni
-        {
-            get { return _vEncIni; }
-            set { _vEncIni = value.Arredondar(3); }
-        }
+        public string dVal { get; set; }
 
         /// <summary>
-        /// LA16 - Valor do Encerrante no final do abastecimento
+        ///     K06 - Preço máximo consumidor
         /// </summary>
-        public decimal vEncFin
+        public decimal vPMC
         {
-            get { return _vEncFin; }
-            set { _vEncFin = value.Arredondar(3); }
+            get { return _vPmc; }
+            set { _vPmc = value.Arredondar(2); }
         }
     }
 }
